Makes synth checks a defined array; adds mean, spread and count; fixes double values; adds notes about synths to install.conf

+++ src/data.h
@@ -70,7 +70,7 @@ struct data_hash_entry
 
 
 DHASH *data_locate( char *path, int len, int adder );
-void data_point_adder( char *path, int len, uint64_t val );
+void data_point_adder( char *path, int len, double val );
 
 line_fn data_line_data;
 line_fn data_line_statsd;
